Revert "Doing upgrade check on admin init and removing admin check in…

+++ lift-core.php
@@ -59,7 +59,7 @@ class Lift_Search {
 			$admin->init();
 		}
 
-		add_action( 'admin_init', array( __CLASS__, '_upgrade_check' ) );
+		add_action( 'init', array( __CLASS__, '_upgrade_check' ) );
 
 		//need cron hooks to be set prior to init
 		add_action( Lift_Batch_Handler::BATCH_CRON_HOOK, array( 'Lift_Batch_Handler', 'send_next_batch' ) );
@@ -489,55 +489,60 @@ class Lift_Search {
 	}
 
 	public static function _upgrade_check() {
-		$current_db_version = get_option( 'lift_db_version', 0 );
-		$queue_all = false;
-		$changed_schema_fields = array( );
 
-		if ( $current_db_version < 2 ) {
-			//queue storage changes
-			$lift_storage_posts = new WP_Query( array(
-				'post_type' => Lift_Document_Update_Queue::STORAGE_POST_TYPE,
-				'fields'    => 'ids'
-			) );
+		if ( is_admin() ) {
 
-			$queue_id = Lift_Document_Update_Queue::get_active_queue_id();
+			$current_db_version = get_option( 'lift_db_version', 0 );
+			$queue_all = false;
+			$changed_schema_fields = array( );
 
-			if ( $lift_storage_posts->have_posts() ) {
-				foreach ( $lift_storage_posts->posts as $post_id ) {
-					if ( $update_meta = get_post_meta( $post_id, 'lift_content', true ) ) {
-						if ( is_string( $update_meta ) )
-							$update_meta = maybe_unserialize( $update_meta ); //previous versions double serialized meta
+			if ( $current_db_version < 2 ) {
+				//queue storage changes
+				$lift_storage_posts = new WP_Query( array(
+					'post_type' => Lift_Document_Update_Queue::STORAGE_POST_TYPE,
+					'fields'    => 'ids'
+				) );
 
-						$meta_key = 'lift_update_' . $update_meta['document_type'] . '_' . $update_meta['document_id'];
-						$new_meta = array(
-							'document_id' => $update_meta['document_id'],
-							'document_type' => $update_meta['document_type'],
-							'action' => $update_meta['action'],
-							'fields' => $update_meta['fields'],
-							'update_date_gmt' => get_post_time( 'Y-m-d H:i:s', true, $post_id ),
-							'update_date' => get_post_time( 'Y-m-d H:i:s', false, $post_id )
-						);
-						update_post_meta( $queue_id, $meta_key, $new_meta );
+				$queue_id = Lift_Document_Update_Queue::get_active_queue_id();
 
-						wp_delete_post( $post_id );
+				if ( $lift_storage_posts->have_posts() ) {
+					foreach ( $lift_storage_posts->posts as $post_id ) {
+						if ( $update_meta = get_post_meta( $post_id, 'lift_content', true ) ) {
+							if ( is_string( $update_meta ) )
+								$update_meta = maybe_unserialize( $update_meta ); //previous versions double serialized meta
+
+							$meta_key = 'lift_update_' . $update_meta['document_type'] . '_' . $update_meta['document_id'];
+							$new_meta = array(
+								'document_id' => $update_meta['document_id'],
+								'document_type' => $update_meta['document_type'],
+								'action' => $update_meta['action'],
+								'fields' => $update_meta['fields'],
+								'update_date_gmt' => get_post_time( 'Y-m-d H:i:s', true, $post_id ),
+								'update_date' => get_post_time( 'Y-m-d H:i:s', false, $post_id )
+							);
+							update_post_meta( $queue_id, $meta_key, $new_meta );
+
+							wp_delete_post( $post_id );
+						}
 					}
 				}
+
+				update_option( 'lift_db_version', 2 );
 			}
 
-			update_option( 'lift_db_version', 2 );
-		}
+			if ( $current_db_version < 4 && self::get_search_domain_name() ) {
+				//schema changes
+				self::update_schema();
 
-		if ( $current_db_version < 4 && self::get_search_domain_name() ) {
-			//schema changes
-			self::update_schema();
+				update_option( 'lift_db_version', 4 );
+			}
 
-			update_option( 'lift_db_version', 4 );
-		}
+			if ( $current_db_version < 5 ) {
+				wp_clear_scheduled_hook( 'lift_index_documents' );
+				wp_clear_scheduled_hook( 'lift_set_endpoints' );
+				update_option( 'lift_db_version', 5 );
+			}
 
-		if ( $current_db_version < 5 ) {
-			wp_clear_scheduled_hook( 'lift_index_documents' );
-			wp_clear_scheduled_hook( 'lift_set_endpoints' );
-			update_option( 'lift_db_version', 5 );
 		}
 
 	}
